requirements.txt
python-dotenv
flask
Flask-SQLAlchemy
flask-login
psycopg2-binary==2.9.3
flask_httpauth
werkzeug
Flask-Migrate
Flask-Testing
requests
gunicorn
uwsgi